初步完成断点续传

+++ qiqu/controller/gm_download_novel_manger.py
@@ -2,13 +2,13 @@
 # -*- coding: utf-8 -*-
 
 import os
-import threading
 import shutil
 import re
 # from multiprocessing import Process
 
 from .gm_crawl_web_data_manger import GMCrawlWebDataManger
-from ..tool import GMFileManger, GMPath, GMJson, GMKey
+from ..tool import GMJson, GMKey, GMThreading
+from ..tool import GMFileManger, GMPath, GMDownloadCache
 from ..model import GMBookChapter, GMBookInfo
 from ..model import GMDownloadStatus, GMDownloadResponse
 from ..model import GMDownloadRequest
@@ -38,20 +38,25 @@ class GMDownloadNovelManger(object):
         manger = GMDownloadNovelManger()
         print("len(manger.__download_map) = " +
               str(len(manger.__download_map)))
-        response = GMDownloadResponse(request, callback)
+
+        msg = None
         if not request or not request.request_id:
-            response.call(GMDownloadStatus.download_error, "任务id出错")
+            msg = "任务id出错"
         elif request.request_id in self.__download_map:
             # 任务列表中存在
-            response.call(GMDownloadStatus.download_error, "正在下载中。。。")
+            msg = "正在下载中。。。"
         elif len(manger.__download_map) >= manger.__max_count:
             # 任务达到最大数
-            response.call(GMDownloadStatus.download_error, "下载数量达到最大！")
-        else:
+            msg = "下载数量达到最大！"
+
+        if not msg:
             # 开始任务
             task = GMDownloadNovelTask(manger.callback_manager)
             manger.__download_map[request.request_id] = task
             task.start_download(request, callback)
+        else:
+            GMDownloadResponse(request, callback).set_params(
+                GMDownloadStatus.download_error, msg).call()
 
     def callback_manager(self, response: GMDownloadResponse):
         if response.code != GMDownloadStatus.downloading_chapter:
@@ -75,30 +80,20 @@ class GMDownloadNovelTask(object):
         """
         self.response_data = {}
         self.response = GMDownloadResponse(request, callback)
-        t = threading.Thread(
-            target=self.__download_novel_with_list_style,
-            kwargs={"request": request},
-            name="__download_novel_with_list_style" + request.request_id)
-        t.start()
-
-    def download_new_book(self, path, book_id, book_name):
-        file_name = self.__download_info_file_name(book_id, book_name)
-        info_path = GMPath.downloadTempFilePath(file_name, '.txt')
-        if os.path.exists(info_path):
-            os.remove(info_path)
+        GMThreading.start(
+            self.__download_novel_with_list_style,
+            "download_" + request.request_id,
+            request=request)
 
     def __callback(self,
                    code=GMDownloadStatus.download_error,
-                   msg: str = None,
-                   progress: str = None):
-        self.response_data["progress"] = progress
+                   msg: str = "",
+                   add_msg: str = ""):
+        self.response.set_params(code,
+                                 GMDownloadResponse.hidden_msg(msg, add_msg),
+                                 self.response_data)
         if self.callback_manager:
-            self.response.code = code
-            self.response.msg = msg
-            self.response.data = self.response_data
             self.callback_manager(self.response)
-        else:
-            self.response.call(code, msg, self.response_data)
 
     def __download_info_file_name(self, book_id, book_name):
         return book_name + "_" + book_id + ""
@@ -108,10 +103,17 @@ class GMDownloadNovelTask(object):
         url = request.url
         book_id = request.book_id
 
-        self.response_data["name"] = GMValue.valueStirng(request.extra, "name")
+        book_name = GMValue.valueStirng(request.extra, "name")
+        last_chapter_id = GMValue.valueStirng(request.extra, "chapter_id")
+
+        self.response_data["name"] = book_name
 
         # 开始下载章节
-        self.__callback(GMDownloadStatus.downloading_chapter, "获取小说信息...")
+        self.__callback(GMDownloadStatus.downloading_chapter,
+                        book_name + "_获取信息中...")
+        # 描述文件更替 待开启
+        if not GMDownloadCache.is_exists(book_id):
+            GMDownloadCache.save(book_id, "", book_name)
 
         # 或取消说首页内容
         gmJsonModel: GMJson = GMCrawlWebDataManger.getNovelListData(
@@ -123,25 +125,45 @@ class GMDownloadNovelTask(object):
 
         if not bookModel or not isinstance(
                 bookModel, GMBookInfo) or not bookModel.chapter_list:
-            self.__callback(GMDownloadStatus.download_error, "获取小说信息失败！！！")
+            self.__callback(GMDownloadStatus.download_error,
+                            book_name + "_获取信息失败！！！")
         else:
             self.response_data["name"] = bookModel.name
-            self.response_data["author"] = bookModel.author
-            self.response_data["url"] = bookModel.url
+            # self.response_data["author"] = bookModel.author
+            # self.response_data["url"] = bookModel.url
             self.response_data["book_id"] = bookModel.book_id
-
+            #warning unfinished code ---
+            bookModel.chapter_list = bookModel.chapter_list[0:20]
             index = 0
             all_count = len(bookModel.chapter_list)
 
             book_name = bookModel.name
             path = ""
-            path_info = ""
 
-            #swarning test code
-            # chapter_list = chapter_list[1:6]
-            self.__callback(GMDownloadStatus.downloading_chapter, "开始下载...")
+            self.__callback(GMDownloadStatus.downloading_chapter,
+                            book_name + "_开始下载...")
 
-            for ele_chapter in bookModel.chapter_list:  # 章节列表
+            chapter_list = list(bookModel.chapter_list)
+
+            is_exists_id = False
+            last_chapter_id_index = 0
+            if last_chapter_id:
+
+                for ele_chapter in chapter_list:  # 章节列表
+                    if last_chapter_id == ele_chapter.chapter_id:
+                        is_exists_id = True
+                        break
+                    last_chapter_id_index += 1
+
+                if is_exists_id:
+                    if last_chapter_id_index + 1 < len(bookModel.chapter_list):
+                        chapter_list = chapter_list[(
+                            last_chapter_id_index +
+                            1):len(bookModel.chapter_list)]
+                    else:
+                        is_exists_id = False
+
+            for ele_chapter in chapter_list:  # 章节列表
 
                 # 获取章节内容
                 chapter_data = GMCrawlWebDataManger.getNovelContentData(
@@ -160,59 +182,47 @@ class GMDownloadNovelTask(object):
                     # 创建路径
                     if len(path) <= 0:
                         path = GMPath.downloadTempFilePath(book_name, '.txt')
-                        path_info = GMPath.downloadTempFilePath(
-                            self.__download_info_file_name(book_id, book_name),
-                            "txt")
-                        if os.path.exists(path) and os.path.exists(path_info):
-                            # 两个文件都存在，继续下载内容 未完成
-                            pass
-                        elif not os.path.exists(path) or not os.path.exists(
-                                path_info):
-                            # 有一个不存在，都删除了重新下载
-                            if os.path.exists(path):
-                                os.remove(path)
-                            if os.path.exists(path_info):
-                                os.remove(path_info)
+                        if last_chapter_id and not is_exists_id\
+                           and os.path.exists(path):
+                            os.remove(path)
 
                     # 追加内容到文本中
                     GMFileManger.appendContent(
                         path, (chapter_title + "\n" + chapterModel.content))
                     # 描述文件更替 待开启
-                    # GMFileManger.replaceContent(path_info,
-                    # chapter_no_download_list)
+                    GMDownloadCache.save(book_id, ele_chapter.chapter_id,
+                                         book_name)
 
                 # 向上层跑出结果
                 re_ret = re.search("第.+?章", chapter_title)
                 if re_ret:
                     chapter_title = re_ret.group()
-                progress = "-".join(
-                    [book_name, chapter_title,
-                     str(index),
-                     str(all_count)])
+                progress = "_".join([
+                    book_name, chapter_title,
+                    str(index + last_chapter_id_index)
+                ])
+                progress += "/" + str(all_count)
                 # 处理打印文案
-                print(index)
                 self.__callback(
-                    GMDownloadStatus.downloading_chapter,
-                    chapter_title + "——下载完成 url：" + ele_chapter.url, progress)
+                    GMDownloadStatus.downloading_chapter, progress,
+                    chapter_title + "——下载完成 url：" + ele_chapter.url)
                 index += 1
 
             # 下载完成
-            # 删除记录文件
-            if os.path.exists(path_info):
-                os.remove(path_info)
+            # 下载完成移动为指导download文件夹下
+            down_file_path = GMPath.downloadFilePath(book_name, '.txt')
 
-                # 下载完成移动为指导download文件夹下
-                down_file_path = GMPath.downloadFilePath(book_name, '.txt')
+            if os.path.exists(down_file_path):
+                os.remove(down_file_path)
 
-                if os.path.exists(down_file_path):
-                    os.remove(down_file_path)
+            if os.path.exists(path):
+                shutil.move(path, GMPath.downloadFilePath())
+            else:
+                print("下载文件不存在 移动到对应位置 失败")
 
-                if os.path.exists(path):
-                    shutil.move(path, GMPath.downloadFilePath())
-                else:
-                    print("下载文件不存在 移动到对应位置 失败")
-
-                print("下载完成， 移动到对应位置 成功")
+            print("下载完成， 移动到对应位置 成功")
+            # 移除缓存文件
+            GMDownloadCache.remove(book_id)
 
             self.__callback(GMDownloadStatus.download_success,
                             book_name + "全书下载完成")
